requirements.txt
asgiref==3.2.3
Babel==2.8.0
Django==3.0.1
django-address==0.2.1
django-crispy-forms==1.8.1
django-phonenumber-field==4.0.0
phonenumbers==8.11.1
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
